fix(ai): handled docx load fail in ai chat
