+++ packages/plugins/@nocobase/plugin-ai/src/server/document-loader/loader.ts
@@ -8,15 +8,12 @@
  */
 
 import PluginFileManagerServer from '@nocobase/plugin-file-manager';
-import { PPTXLoader } from '@langchain/community/document_loaders/fs/pptx';
-import { DocxLoader } from '@langchain/community/document_loaders/fs/docx';
 import { Document } from '@langchain/core/documents';
 import { Readable } from 'node:stream';
 import { Worker } from 'node:worker_threads';
 import { SUPPORTED_DOCUMENT_EXTNAMES } from './constants';
 import { ParseableFile, SupportedDocumentExtname } from './types';
 import path from 'node:path';
-import { TextLoader } from './vendor/langchain/document_loaders/fs/text';
 import { resolveExtname } from './utils';
 
 export class DocumentLoader {
@@ -30,22 +27,7 @@ export class DocumentLoader {
 
     const { stream, contentType } = await this.fileManager.getFileStream(file as any);
     const blob = await this.streamToBlob(stream, contentType ?? file.mimetype);
-
-    switch (extname) {
-      case '.pdf':
-        return this.loadPdf(blob);
-      case '.ppt':
-      case '.pptx':
-        return this.loadPpt(blob);
-      case '.doc':
-        return this.loadDoc(blob, { type: 'doc' });
-      case '.docx':
-        return this.loadDoc(blob, { type: 'docx' });
-      case '.txt':
-        return this.loadTxt(blob);
-      default:
-        return [];
-    }
+    return await this.loadByWorker(extname, blob);
   }
 
   private async streamToBlob(stream: Readable, mimeType = 'application/octet-stream') {
@@ -59,10 +41,10 @@ export class DocumentLoader {
     return new Blob(chunks, { type: mimeType });
   }
 
-  private async loadPdf(blob: Blob): Promise<Document[]> {
+  private async loadByWorker(extname: SupportedDocumentExtname, blob: Blob): Promise<Document[]> {
     const buffer = Buffer.from(await blob.arrayBuffer());
     const isTsRuntime = __filename.endsWith('.ts');
-    const workerPath = path.join(__dirname, `pdf-loader.worker.${isTsRuntime ? 'ts' : 'js'}`);
+    const workerPath = path.join(__dirname, `loader.worker.${isTsRuntime ? 'ts' : 'js'}`);
     const worker = new Worker(workerPath, {
       execArgv: isTsRuntime ? ['--require', 'tsx/cjs'] : undefined,
     });
@@ -90,28 +72,17 @@ export class DocumentLoader {
       worker.once('error', (error) => close(error));
       worker.once('exit', (code) => {
         if (!settled && code !== 0) {
-          close(new Error(`PDF worker exited with code ${code}`));
+          close(new Error(`Document loader worker exited with code ${code}`));
         }
       });
 
-      worker.postMessage({ buffer: Uint8Array.from(buffer) });
+      worker.postMessage({
+        extname,
+        mimeType: blob.type,
+        buffer: Uint8Array.from(buffer),
+      });
     }).finally(() => {
       worker.terminate().catch(() => undefined);
     });
-  }
-
-  private async loadDoc(blob: Blob, options: { type: 'docx' | 'doc' }): Promise<Document[]> {
-    const loader = new DocxLoader(blob, options);
-    return loader.load();
-  }
-
-  private async loadPpt(blob: Blob): Promise<Document[]> {
-    const loader = new PPTXLoader(blob);
-    return loader.load();
-  }
-
-  private async loadTxt(blob: Blob): Promise<Document[]> {
-    const loader = new TextLoader(blob);
-    return loader.load();
   }
 }
