+++ packages/plugins/@nocobase/plugin-ai/src/server/document-loader/types.ts
@@ -41,6 +41,7 @@ export type ParseableFile = {
   mimetype?: string;
   storageId?: number;
   meta?: Record<string, any>;
+  size?: number;
 };
 
 export type DocumentLoaderLike = {
